Split up App into more components.
