+++ src/pages/popup.tsx
@@ -3,7 +3,7 @@ import ReactDom from 'react-dom';
 import { Provider } from 'react-redux';
 import { Store } from 'webext-redux';
 
-import { App } from '../containers/App';
+import { App } from '../components/App';
 
 const store = new Store();
